gaz-server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-multipart>=0.0.9
python-dotenv>=1.0
httpx>=0.27
openai>=1.30
groq>=0.9
supabase>=2.4
qdrant-client>=1.9
yt-dlp>=2024.3.10
youtube-transcript-api>=0.6.2
ffmpeg-python>=0.2.0
langchain-text-splitters>=0.0.1
tiktoken>=0.7
